backend/pyproject.toml
[project]
name = "friendos"
version = "0.1.0"
description = "Personal CRM for tracking friends"
requires-python = ">=3.13"
dependencies = [
    "pydantic-settings>=2.7.0",
    "email-validator>=2.2.0",
    "psycopg[binary]>=3.2.0",
    "sqlalchemy>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.20",
    "alembic>=1.14.0",
    "httpx>=0.28.0",
    "bcrypt>=4.2.0",
    "python-jose[cryptography]>=3.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
